+++ pocketbase_helpers/example/bin/example.dart
@@ -1,8 +1,0 @@
-import 'package:example/example.dart' as example;
-
-void main(List<String> arguments) async {
-  print('Retrieving users...');
-  final users = await example.fetch();
-  print('Users retrieved!');
-  print(users);
-}