+++ pocketbase_helpers/lib/src/io_only/io.dart
@@ -1,13 +1,16 @@
 import 'dart:io';
 import 'dart:typed_data';
 
-Map<String, Uint8List> pathsToFiles(List<String> filePaths) {
-  final fileMap = <String, Uint8List>{};
-  for (final path in filePaths) {
-    final file = File(path);
-    final name = file.uri.pathSegments.last;
-    fileMap[name] = file.readAsBytesSync();
-  }
+Future<Map<String, Uint8List>> pathsToFiles(List<String> filePaths) async {
+  final data = await [for (final path in filePaths) pathToFile(path)].wait;
 
-  return fileMap;
+  return {for (final e in data) e.$1: e.$2};
+}
+
+Future<(String, Uint8List)> pathToFile(String path) async {
+  final file = File(path);
+  final name = file.uri.pathSegments.last;
+  final bytes = await file.readAsBytes();
+
+  return (name, bytes);
 }