+++ pocketbase_helpers/lib/src/io_only/stub.dart
@@ -1,5 +1,9 @@
 import 'dart:typed_data';
 
-Map<String, Uint8List> pathsToFiles(List<String> filePaths) {
+Future<Map<String, Uint8List>> pathsToFiles(List<String> filePaths) {
+  throw UnsupportedError('The method pathsToFiles() is not available for web');
+}
+
+Future<(String, Uint8List)> pathToFile(String path) {
   throw UnsupportedError('The method pathsToFiles() is not available for web');
 }